Refactor - Move test_register_error in pylint.message test
